refactor: Add ManageQuickActions.dart
