+++ lib/main.dart
@@ -11,6 +11,7 @@ import 'package:locus/app_wrappers/CheckViewAlarmsLive.dart';
 import 'package:locus/app_wrappers/HandleNotifications.dart';
 import 'package:locus/app_wrappers/LocationHistoryUpdater.dart';
 import 'package:locus/app_wrappers/RegisterBackgroundListeners.dart';
+import 'package:locus/app_wrappers/ManageQuickActions.dart';
 import 'package:locus/app_wrappers/UniLinksHandler.dart';
 import 'package:locus/app_wrappers/UpdateLastLocationToSettings.dart';
 import 'package:locus/app_wrappers/UpdateLocaleToSettings.dart';
@@ -27,7 +28,7 @@ import 'package:provider/provider.dart';
 const storage = FlutterSecureStorage();
 
 final StreamController<NotificationResponse> selectedNotificationsStream =
-    StreamController.broadcast();
+StreamController.broadcast();
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -55,7 +56,7 @@ void main() async {
   );
 
   FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-      FlutterLocalNotificationsPlugin();
+  FlutterLocalNotificationsPlugin();
   const initializationSettings = InitializationSettings(
     android: AndroidInitializationSettings("ic_launcher_foreground"),
     iOS: DarwinInitializationSettings(),
@@ -110,6 +111,7 @@ void main() async {
           UpdateLocaleToSettings(),
           HandleNotifications(),
           CheckViewAlarmsLive(),
+          ManageQuickActions(),
           App(),
         ],
       ),
