+++ lib/screens/ShortcutScreen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+import 'package:locus/app_wrappers/ManageQuickActions.dart';
 import 'package:locus/constants/spacing.dart';
-import 'package:locus/init_quick_actions.dart';
 import 'package:locus/services/task_service/index.dart';
 import 'package:locus/services/timers_service.dart';
 import 'package:locus/utils/location/index.dart';
